+++ gobblin-service/src/test/java/org/apache/gobblin/service/modules/orchestration/MysqlUserQuotaManagerTest.java
@@ -19,6 +19,8 @@ package org.apache.gobblin.service.modules.orchestration;
 
 import java.io.IOException;
 
+import java.sql.Connection;
+import java.sql.SQLException;
 import org.testng.Assert;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
@@ -54,53 +56,67 @@ public class MysqlUserQuotaManagerTest {
   @Test
   public void testRunningDagStore() throws Exception {
     String dagId = DagManagerUtils.generateDagId(DagManagerTest.buildDag("dagId", 1234L, "", 1).getNodes().get(0)).toString();
+    Connection connection = this.quotaManager.quotaStore.dataSource.getConnection();
     Assert.assertFalse(this.quotaManager.containsDagId(dagId));
-    this.quotaManager.addDagId(dagId);
+    this.quotaManager.addDagId(connection, dagId);
+    connection.commit();
     Assert.assertTrue(this.quotaManager.containsDagId(dagId));
-    Assert.assertTrue(this.quotaManager.removeDagId(dagId));
+    Assert.assertTrue(this.quotaManager.removeDagId(connection, dagId));
+    connection.commit();
     Assert.assertFalse(this.quotaManager.containsDagId(dagId));
-    Assert.assertFalse(this.quotaManager.removeDagId(dagId));
+    Assert.assertFalse(this.quotaManager.removeDagId(connection, dagId));
+    connection.commit();
+    connection.close();
   }
 
     @Test
   public void testIncreaseCount() throws Exception {
-    int prevCount = this.quotaManager.incrementJobCount(PROXY_USER, AbstractUserQuotaManager.CountType.USER_COUNT);
+    Connection connection = this.quotaManager.quotaStore.dataSource.getConnection();
+    int prevCount = this.quotaManager.incrementJobCount(connection, PROXY_USER, AbstractUserQuotaManager.CountType.USER_COUNT);
     Assert.assertEquals(prevCount, 0);
 
-    prevCount = this.quotaManager.incrementJobCount(PROXY_USER, AbstractUserQuotaManager.CountType.USER_COUNT);
+    prevCount = this.quotaManager.incrementJobCount(connection, PROXY_USER, AbstractUserQuotaManager.CountType.USER_COUNT);
     Assert.assertEquals(prevCount, 1);
     Assert.assertEquals(this.quotaManager.getCount(PROXY_USER, AbstractUserQuotaManager.CountType.USER_COUNT), 2);
 
-    prevCount = this.quotaManager.incrementJobCount(PROXY_USER, AbstractUserQuotaManager.CountType.FLOWGROUP_COUNT);
+    prevCount = this.quotaManager.incrementJobCount(connection, PROXY_USER, AbstractUserQuotaManager.CountType.FLOWGROUP_COUNT);
     Assert.assertEquals(prevCount, 0);
 
-    prevCount = this.quotaManager.incrementJobCount(PROXY_USER, AbstractUserQuotaManager.CountType.FLOWGROUP_COUNT);
+    prevCount = this.quotaManager.incrementJobCount(connection, PROXY_USER, AbstractUserQuotaManager.CountType.FLOWGROUP_COUNT);
     Assert.assertEquals(prevCount, 1);
+    connection.close();
   }
 
   @Test(dependsOnMethods = "testIncreaseCount")
   public void testDecreaseCount() throws Exception {
-    this.quotaManager.decrementJobCount(PROXY_USER, AbstractUserQuotaManager.CountType.USER_COUNT);
+    Connection connection = this.quotaManager.quotaStore.dataSource.getConnection();
+    this.quotaManager.decrementJobCount(connection, PROXY_USER, AbstractUserQuotaManager.CountType.USER_COUNT);
     Assert.assertEquals(this.quotaManager.getCount(PROXY_USER, AbstractUserQuotaManager.CountType.USER_COUNT), 1);
 
-    this.quotaManager.decrementJobCount(PROXY_USER, AbstractUserQuotaManager.CountType.USER_COUNT);
+    this.quotaManager.decrementJobCount(connection, PROXY_USER, AbstractUserQuotaManager.CountType.USER_COUNT);
     Assert.assertEquals(this.quotaManager.getCount(PROXY_USER, AbstractUserQuotaManager.CountType.USER_COUNT), 0);
 
-    this.quotaManager.decrementJobCount(PROXY_USER, AbstractUserQuotaManager.CountType.USER_COUNT);
+    this.quotaManager.decrementJobCount(connection, PROXY_USER, AbstractUserQuotaManager.CountType.USER_COUNT);
     Assert.assertEquals(this.quotaManager.getCount(PROXY_USER, AbstractUserQuotaManager.CountType.USER_COUNT), 0);
 
-    this.quotaManager.decrementJobCount(PROXY_USER, AbstractUserQuotaManager.CountType.FLOWGROUP_COUNT);
+    this.quotaManager.decrementJobCount(connection, PROXY_USER, AbstractUserQuotaManager.CountType.FLOWGROUP_COUNT);
     Assert.assertEquals(this.quotaManager.getCount(PROXY_USER, AbstractUserQuotaManager.CountType.FLOWGROUP_COUNT), 1);
-    this.quotaManager.decrementJobCount(PROXY_USER, AbstractUserQuotaManager.CountType.FLOWGROUP_COUNT);
+    this.quotaManager.decrementJobCount(connection, PROXY_USER, AbstractUserQuotaManager.CountType.FLOWGROUP_COUNT);
     // on count reduced to zero, the row should get deleted and the get call should return -1 instead of 0.
     Assert.assertEquals(this.quotaManager.getCount(PROXY_USER, AbstractUserQuotaManager.CountType.FLOWGROUP_COUNT), -1);
   }
 
   class ChangeCountRunnable implements Runnable {
     boolean increaseOrDecrease;
+    Connection connection;
 
     public ChangeCountRunnable(boolean increaseOrDecrease) {
       this.increaseOrDecrease = increaseOrDecrease;
+      try {
+        this.connection = MysqlUserQuotaManagerTest.this.quotaManager.quotaStore.dataSource.getConnection();
+      } catch (Exception e) {
+        throw new RuntimeException(e);
+      }
     }
 
     @Override
@@ -109,11 +125,11 @@ public class MysqlUserQuotaManagerTest {
       while (i++ < 1000) {
         try {
           if (increaseOrDecrease) {
-            MysqlUserQuotaManagerTest.this.quotaManager.incrementJobCount(PROXY_USER, AbstractUserQuotaManager.CountType.USER_COUNT);
+            MysqlUserQuotaManagerTest.this.quotaManager.incrementJobCount(this.connection, PROXY_USER, AbstractUserQuotaManager.CountType.USER_COUNT);
           } else {
-            MysqlUserQuotaManagerTest.this.quotaManager.decrementJobCount(PROXY_USER, AbstractUserQuotaManager.CountType.USER_COUNT);
+            MysqlUserQuotaManagerTest.this.quotaManager.decrementJobCount(this.connection, PROXY_USER, AbstractUserQuotaManager.CountType.USER_COUNT);
           }
-        } catch (IOException e) {
+        } catch (IOException | SQLException e) {
           Assert.fail("Thread got an exception.", e);
         }
       }
